intellinote

+++ backend/folder/folder.controller.php
@@ -2,16 +2,18 @@
 
 if(!defined ('CONSTANT') )
     die ('accès interdit') ;
-    
+
+require_once('generique/generique.controller.php');
 require_once('folder/folder.modele.php');
 
-    class controllerFolder {
+    class controllerFolder extends ControllerGenerique {
 
         // conn
         private $modeleFolder;
 
         public function __construct () {
-            //parent::__construct();// on lui donne accès aux méthodes de la class controleur generique
+            parent::__construct();
+
             $this->modeleFolder = new modeleFolder();
             
             $data = json_decode(file_get_contents("php://input"));
@@ -34,54 +36,46 @@ require_once('folder/folder.modele.php');
 
         // CREATE Folder
         public function createFolder(){
-            if($this->modeleFolder->createFolder()){
-                echo json_encode("folder was created");
-            } else{
-                http_response_code(500);
-                echo json_encode(
-                    array("message" => "Folder was not created.")
-                );
+            $decoded = self::verifyToken();
+            if($decoded && $decoded->data->idUser){
+                if($this->modeleFolder->createFolder()){
+                    echo json_encode("folder was created");
+                } else{
+                    http_response_code(500);
+                    echo json_encode(
+                        array("message" => "Folder was not created.")
+                    );
+                }
             }
         }
-
-        // GET Folder
-        public function singnIn(){
-            $dataRow = $this->modeleFolder->getSingleFolder();
-            if($dataRow){
-                $folder_Arr = array(
-                    "idFolder" => $dataRow['idFolder'],
-                    "name" => $dataRow['name'],
-                    "id_User" => $dataRow['id_User'],
-                    "password" => $dataRow['password']
-                );
-                
-                http_response_code(200);
-                echo json_encode($folder_Arr);
-
-            }
-            else{
-                http_response_code(404);
-                echo json_encode("Folder record not found.");
-            }
-                
-        }        
+ 
 
         // UPDATE Folder
         public function updateFolder(){
-            if($this->modeleFolder->updateFolder()){
-                echo json_encode("Folder record updated.");
+            $decoded = self::verifyToken();
+            $datarow = $this->modeleFolder->getSingleFolder();
+            //il faut verifier en plus si le dossier lui appartient
+            if($datarow && $decoded && $decoded->data->idUser == $datarow["id_User"]){
+                if($this->modeleFolder->updateFolder()){
+                    echo json_encode("Folder record updated.");
+                }
+                else{
+                    http_response_code(500);
+                    echo json_encode(
+                        array("message" => "Folder record could not be updated.")
+                    );
+                }
             }
             else{
-                http_response_code(500);
-                echo json_encode(
-                    array("message" => "Folder record could not be updated.")
-                );
+                http_response_code(404);
+                echo json_encode(array("message" => "Credentials not found"));
             }
         }
 
       
         // GET Folders
         public function getFolder(){
+
             if(isset($_GET['idFolder']) ){
                 $this->modeleFolder->idFolder = $_GET['idFolder']; 
             }
@@ -94,46 +88,52 @@ require_once('folder/folder.modele.php');
             if(isset($_GET['id_Folder']) ){
                 $this->modeleFolder->id_Folder = $_GET['id_Folder'];
             }
-                    
-            $stmt =$this->modeleFolder->findFolderByParams();
-            if($stmt){
-                $itemCount = $stmt->rowCount();
+            
+            $decoded = self::verifyToken();
+            if($decoded && $decoded->data->idUser){
+                $stmt =$this->modeleFolder->findFolderByParams();
+                if($stmt){
+                    $itemCount = $stmt->rowCount();
 
-                $folderArr = array();
+                    $folderArr = array();
 
-                while ($row = $stmt->fetch(PDO::FETCH_ASSOC)){
-                    extract($row);
-                    $e = array(
-                        "idFolder" => $idFolder,
-                        "name" => $name,
-                        "id_User" => $id_User,
-                        "id_Folder" => $id_Folder
-                    );
+                    while ($row = $stmt->fetch(PDO::FETCH_ASSOC)){
+                        extract($row);
+                        $e = array(
+                            "idFolder" => $idFolder,
+                            "name" => $name,
+                            "id_User" => $id_User,
+                            "id_Folder" => $id_Folder
+                        );
 
-                    array_push($folderArr, $e);
+                        array_push($folderArr, $e);
+                    }
+                    echo json_encode($folderArr);
                 }
-                echo json_encode($folderArr);
-            }
-            else{
-                http_response_code(404);
-                echo json_encode(
-                    array("message" => "Data not found.")
-                );
+                else{
+                    http_response_code(404);
+                    echo json_encode(
+                        array("message" => "Data not found.")
+                    );
+                }
             }
         }
 
         // DELETE Folder
         function deleteFolder(){
-            if(isset($_GET['idFolder']) ){
-                $this->modeleFolder->idFolder = $_GET['idFolder']; 
-            }
-            if($this->modeleFolder->deleteFolder()){
-                echo json_encode("Folder deleted.");
-            } else{
-                http_response_code(500);
-                echo json_encode(
-                    array("message" => "Folder was not deleted")
-                );
+            $decoded = self::verifyToken();
+            if($decoded && $decoded->data->idUser){
+                if(isset($_GET['idFolder']) ){
+                    $this->modeleFolder->idFolder = $_GET['idFolder']; 
+                }
+                if($this->modeleFolder->deleteFolder()){
+                    echo json_encode("Folder deleted.");
+                } else{
+                    http_response_code(500);
+                    echo json_encode(
+                        array("message" => "Folder was not deleted")
+                    );
+                }
             }
         }
